# control-plane/openapi.json
{
  "components": {
    "schemas": {
...
      "AccountUsageResponse": {
        "properties": {
          "events": {
            "title": "Events",
            "type": "integer"
          },
          "first_occurred_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Occurred At"
          },
          "input_tokens": {
            "title": "Input Tokens",
            "type": "integer"
          },
          "last_occurred_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Occurred At"
          },
          "output_tokens": {
            "title": "Output Tokens",
            "type": "integer"
          }
        },
        "required": [
          "events",
          "input_tokens",
          "output_tokens",
          "first_occurred_at",
          "last_occurred_at"
        ],
        "title": "AccountUsageResponse",
        "type": "object"
      },
...
      "DeploymentCreateRequest": {
        "properties": {
          "model_alias": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Model Alias",
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "pattern": "^[a-z0-9]([a-z0-9-]{0,198}[a-z0-9])?$",
            "title": "Name",
            "type": "string"
          },
          "spec": {
            "$ref": "#/components/schemas/DeploymentSpec"
          }
        },
        "required": [
          "name",
          "model_alias",
          "spec"
        ],
        "title": "DeploymentCreateRequest",
        "type": "object"
      },
      "DeploymentResponse": {
        "properties": {
          "account_id": {
            "format": "uuid",
            "title": "Account Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "desired_spec": {
            "title": "Desired Spec",
            "type": "object"
          },
          "generation": {
            "title": "Generation",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "model_alias": {
            "title": "Model Alias",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "account_id",
          "name",
          "model_alias",
          "desired_spec",
          "generation",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "DeploymentResponse",
        "type": "object"
      },
...
      "DeploymentUpdateRequest": {
        "properties": {
          "spec": {
            "$ref": "#/components/schemas/DeploymentSpec"
          }
        },
        "required": [
          "spec"
        ],
        "title": "DeploymentUpdateRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlacementCreateRequest": {
        "description": "Where to place a deployment, or nothing to let the platform choose.\n\nOmitting ``stamp_id`` is what makes the platform managed rather than manual: the control\nplane filters its stamps by entitlement, region, liveness and fit and picks the least\nloaded one (ADR 0013). Naming a stamp still checks that it fits, and refuses with the\nreason if it does not.",
        "properties": {
          "region": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "stamp_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stamp Id"
          }
        },
        "title": "PlacementCreateRequest",
        "type": "object"
      },
      "PlacementResponse": {
        "properties": {
          "account_id": {
            "format": "uuid",
            "title": "Account Id",
            "type": "string"
          },
          "deployment_id": {
            "format": "uuid",
            "title": "Deployment Id",
            "type": "string"
          },
          "desired_generation": {
            "title": "Desired Generation",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "observed_generation": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed Generation"
          },
          "stamp_id": {
            "format": "uuid",
            "title": "Stamp Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "account_id",
          "deployment_id",
          "stamp_id",
          "desired_generation",
          "observed_generation",
          "status"
        ],
        "title": "PlacementResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "Accounts, credentials, deployment intent, and inference stamps.",
    "title": "Fabric Control Plane",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/accounts/{account_id}/deployments": {
      "get": {
        "operationId": "list_all_v1_accounts__account_id__deployments_get",
        "parameters": [
          {
            "description": "Account identifier",
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "description": "Account identifier",
              "format": "uuid",
              "title": "Account Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DeploymentResponse"
                  },
                  "title": "Response List All V1 Accounts  Account Id  Deployments Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List deployments",
        "tags": [
          "deployments"
        ]
      },
      "post": {
        "description": "Create deployment intent, once per idempotency key.\n\nA caller that loses the response to this request has no safe move without a key:\nretrying may create a second deployment, and not retrying may leave none. With a\nkey, the retry returns the first deployment instead of creating another.",
        "operationId": "create_v1_accounts__account_id__deployments_post",
        "parameters": [
          {
            "description": "Account identifier",
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "description": "Account identifier",
              "format": "uuid",
              "title": "Account Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeploymentCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create deployment intent",
        "tags": [
          "deployments"
        ]
      }
    },
    "/v1/accounts/{account_id}/deployments/usage": {
      "get": {
        "description": "Aggregate operational usage reported across all account deployments.",
        "operationId": "read_account_usage_v1_accounts__account_id__deployments_usage_get",
        "parameters": [
          {
            "description": "Account identifier",
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "description": "Account identifier",
              "format": "uuid",
              "title": "Account Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountUsageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read account-wide reported usage totals",
        "tags": [
          "deployments"
        ]
      }
    },
    "/v1/accounts/{account_id}/deployments/{deployment_id}": {
      "delete": {
        "operationId": "delete_v1_accounts__account_id__deployments__deployment_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "deployment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Deployment Id",
              "type": "string"
            }
          },
          {
            "description": "Account identifier",
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "description": "Account identifier",
              "format": "uuid",
              "title": "Account Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete deployment intent",
        "tags": [
          "deployments"
        ]
      },
      "get": {
        "operationId": "read_one_v1_accounts__account_id__deployments__deployment_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "deployment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Deployment Id",
              "type": "string"
            }
          },
          {
            "description": "Account identifier",
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "description": "Account identifier",
              "format": "uuid",
              "title": "Account Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read one deployment",
        "tags": [
          "deployments"
        ]
      },
      "patch": {
        "operationId": "update_v1_accounts__account_id__deployments__deployment_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "deployment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Deployment Id",
              "type": "string"
            }
          },
          {
            "description": "Account identifier",
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "description": "Account identifier",
              "format": "uuid",
              "title": "Account Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeploymentUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replace the desired spec and advance the generation",
        "tags": [
          "deployments"
        ]
      }
    },
    "/v1/accounts/{account_id}/deployments/{deployment_id}/placements": {
      "get": {
        "operationId": "read_placements_v1_accounts__account_id__deployments__deployment_id__placements_get",
        "parameters": [
          {
            "in": "path",
            "name": "deployment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Deployment Id",
              "type": "string"
            }
          },
          {
            "description": "Account identifier",
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "description": "Account identifier",
              "format": "uuid",
              "title": "Account Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlacementResponse"
                  },
                  "title": "Response Read Placements V1 Accounts  Account Id  Deployments  Deployment Id  Placements Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List placements",
        "tags": [
          "deployments"
        ]
      },
      "post": {
        "description": "Place a deployment, refusing what the stamp cannot hold.\n\nWith no ``stamp_id`` the platform chooses one that fits. With one, it is authorized and\nthen checked: an assignment a stamp cannot serve used to answer ``201`` and then never\nstart (ADR 0013). The chosen stamp is on the response either way.",
        "operationId": "assign_placement_v1_accounts__account_id__deployments__deployment_id__placements_post",
        "parameters": [
          {
            "in": "path",
            "name": "deployment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Deployment Id",
              "type": "string"
            }
          },
          {
            "description": "Account identifier",
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "description": "Account identifier",
              "format": "uuid",
              "title": "Account Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/PlacementCreateRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlacementResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Assign a deployment to a stamp that fits, or pick one",
        "tags": [
          "deployments"
        ]
      }
    },
    "/v1/accounts/{account_id}/deployments/{deployment_id}/status": {
      "get": {
        "operationId": "read_status_v1_accounts__account_id__deployments__deployment_id__status_get",
        "parameters": [
          {
            "in": "path",
            "name": "deployment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Deployment Id",
              "type": "string"
            }
          },
          {
            "description": "Account identifier",
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "description": "Account identifier",
              "format": "uuid",
              "title": "Account Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DeploymentStatusResponse"
                  },
                  "title": "Response Read Status V1 Accounts  Account Id  Deployments  Deployment Id  Status Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read reported status per stamp",
        "tags": [
          "deployments"
        ]
      }
    },
    "/v1/accounts/{account_id}/deployments/{deployment_id}/usage": {
      "get": {
        "description": "Aggregate usage a collector reported for this deployment.\n\nUsage is operational in MVP, not billing-grade accounting: records are\ndeduplicated per stamp but delivery is at-least-once and a stamp that never\nreports simply contributes nothing.",
        "operationId": "read_usage_v1_accounts__account_id__deployments__deployment_id__usage_get",
        "parameters": [
          {
            "in": "path",
            "name": "deployment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Deployment Id",
              "type": "string"
            }
          },
          {
            "description": "Account identifier",
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "description": "Account identifier",
              "format": "uuid",
              "title": "Account Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentUsageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read reported usage totals",
        "tags": [
          "deployments"
        ]
      }
    },
...
  }
}